Merge

+++ xen/arch/x86/mm.c
@@ -106,7 +106,7 @@
 #ifdef VERBOSE
 #define MEM_LOG(_f, _a...)                           \
   printk("DOM%u: (file=mm.c, line=%d) " _f "\n", \
-         current->domain->id , __LINE__ , ## _a )
+         current->domain->domain_id , __LINE__ , ## _a )
 #else
 #define MEM_LOG(_f, _a...) ((void)0)
 #endif
@@ -160,8 +160,11 @@ void __init init_frametable(void)
         p = alloc_boot_pages(min(frame_table_size - i, 4UL << 20), 4UL << 20);
         if ( p == 0 )
             panic("Not enough memory for frame table\n");
-        map_pages(idle_pg_table, FRAMETABLE_VIRT_START + i, p, 
-                  4UL << 20, PAGE_HYPERVISOR);
+        map_pages_to_xen(
+            FRAMETABLE_VIRT_START + i,
+            p >> PAGE_SHIFT,
+            4UL << (20-PAGE_SHIFT),
+            PAGE_HYPERVISOR);
     }
 
     memset(frame_table, 0, frame_table_size);
@@ -171,7 +174,7 @@ void arch_init_memory(void)
 {
     extern void subarch_init_memory(struct domain *);
 
-    unsigned long i, j, pfn, nr_pfns;
+    unsigned long i, pfn, rstart_pfn, rend_pfn;
     struct pfn_info *page;
 
     memset(percpu_info, 0, sizeof(percpu_info));
@@ -183,7 +186,7 @@ void arch_init_memory(void)
      */
     dom_xen = alloc_domain_struct();
     atomic_set(&dom_xen->refcnt, 1);
-    dom_xen->id = DOMID_XEN;
+    dom_xen->domain_id = DOMID_XEN;
 
     /*
      * Initialise our DOMID_IO domain.
@@ -192,7 +195,7 @@ void arch_init_memory(void)
      */
     dom_io = alloc_domain_struct();
     atomic_set(&dom_io->refcnt, 1);
-    dom_io->id = DOMID_IO;
+    dom_io->domain_id = DOMID_IO;
 
     /* First 1MB of RAM is historically marked as I/O. */
     for ( i = 0; i < 0x100; i++ )
@@ -203,25 +206,26 @@ void arch_init_memory(void)
         page_set_owner(page, dom_io);
     }
  
-    /* Any non-RAM areas in the e820 map are considered to be for I/O. */
-    for ( i = 0; i < e820.nr_map; i++ )
+    /* Any areas not specified as RAM by the e820 map are considered I/O. */
+    for ( i = 0, pfn = 0; i < e820.nr_map; i++ )
     {
-        if ( e820.map[i].type == E820_RAM )
+        if ( e820.map[i].type != E820_RAM )
             continue;
-        pfn = e820.map[i].addr >> PAGE_SHIFT;
-        nr_pfns = (e820.map[i].size +
-                   (e820.map[i].addr & ~PAGE_MASK) +
-                   ~PAGE_MASK) >> PAGE_SHIFT;
-        for ( j = 0; j < nr_pfns; j++ )
+        /* Every page from cursor to start of next RAM region is I/O. */
+        rstart_pfn = PFN_UP(e820.map[i].addr);
+        rend_pfn   = PFN_DOWN(e820.map[i].addr + e820.map[i].size);
+        while ( pfn < rstart_pfn )
         {
-            if ( !pfn_valid(pfn+j) )
-                continue;
-            page = &frame_table[pfn+j];
+            BUG_ON(!pfn_valid(pfn));
+            page = &frame_table[pfn++];
             page->count_info        = PGC_allocated | 1;
             page->u.inuse.type_info = PGT_writable_page | PGT_validated | 1;
             page_set_owner(page, dom_io);
         }
+        /* Skip the RAM region. */
+        pfn = rend_pfn;
     }
+    BUG_ON(pfn != max_page);
 
     subarch_init_memory(dom_xen);
 }
@@ -1162,7 +1166,7 @@ void put_page_type(struct pfn_info *page)
          * See domain.c:relinquish_list().
          */
         ASSERT((x & PGT_validated) || 
-               test_bit(DF_DYING, &page_get_owner(page)->flags));
+               test_bit(_DOMF_dying, &page_get_owner(page)->domain_flags));
 
         if ( unlikely((nx & PGT_count_mask) == 0) )
         {
@@ -1392,7 +1396,7 @@ static int set_foreigndom(unsigned int cpu, domid_t domid)
             percpu_info[cpu].foreign = dom_io;
             break;
         default:
-            MEM_LOG("Dom %u cannot set foreign dom\n", d->id);
+            MEM_LOG("Dom %u cannot set foreign dom\n", d->domain_id);
             okay = 0;
             break;
         }
@@ -1645,7 +1649,7 @@ int do_mmuext_op(
             if ( shadow_mode_external(d) )
             {
                 MEM_LOG("ignoring SET_LDT hypercall from external "
-                        "domain %u\n", d->id);
+                        "domain %u\n", d->domain_id);
                 okay = 0;
                 break;
             }
@@ -1676,7 +1680,7 @@ int do_mmuext_op(
         case MMUEXT_REASSIGN_PAGE:
             if ( unlikely(!IS_PRIV(d)) )
             {
-                MEM_LOG("Dom %u has no reassignment priv", d->id);
+                MEM_LOG("Dom %u has no reassignment priv", d->domain_id);
                 okay = 0;
                 break;
             }
@@ -1711,13 +1715,13 @@ int do_mmuext_op(
              * it is dying. 
              */
             ASSERT(e->tot_pages <= e->max_pages);
-            if ( unlikely(test_bit(DF_DYING, &e->flags)) ||
+            if ( unlikely(test_bit(_DOMF_dying, &e->domain_flags)) ||
                  unlikely(e->tot_pages == e->max_pages) ||
                  unlikely(IS_XEN_HEAP_FRAME(page)) )
             {
                 MEM_LOG("Transferee has no reservation headroom (%d,%d), or "
                         "page is in Xen heap (%lx), or dom is dying (%ld).\n",
-                        e->tot_pages, e->max_pages, op.mfn, e->flags);
+                        e->tot_pages, e->max_pages, op.mfn, e->domain_flags);
                 okay = 0;
                 goto reassign_fail;
             }
@@ -1738,7 +1742,7 @@ int do_mmuext_op(
                 {
                     MEM_LOG("Bad page values %lx: ed=%p(%u), sd=%p,"
                             " caf=%08x, taf=%08x\n", page_to_pfn(page),
-                            d, d->id, unpickle_domptr(_nd), x,
+                            d, d->domain_id, unpickle_domptr(_nd), x,
                             page->u.inuse.type_info);
                     okay = 0;
                     goto reassign_fail;
@@ -1990,7 +1994,7 @@ int do_mmu_update(
             {
                 shadow_lock(FOREIGNDOM);
                 printk("privileged guest dom%d requests pfn=%lx to map mfn=%lx for dom%d\n",
-                       d->id, gpfn, mfn, FOREIGNDOM->id);
+                       d->domain_id, gpfn, mfn, FOREIGNDOM->domain_id);
                 set_machinetophys(mfn, gpfn);
                 set_p2m_entry(FOREIGNDOM, gpfn, mfn, &sh_mapcache, &mapcache);
                 okay = 1;
@@ -2244,11 +2248,13 @@ void destroy_gdt(struct exec_domain *ed)
     int i;
     unsigned long pfn;
 
-    for ( i = 0; i < 16; i++ )
+    ed->arch.guest_context.gdt_ents = 0;
+    for ( i = 0; i < FIRST_RESERVED_GDT_PAGE; i++ )
     {
         if ( (pfn = l1e_get_pfn(ed->arch.perdomain_ptes[i])) != 0 )
             put_page_and_type(&frame_table[pfn]);
         ed->arch.perdomain_ptes[i] = l1e_empty();
+        ed->arch.guest_context.gdt_frames[i] = 0;
     }
 }
 
@@ -2259,56 +2265,31 @@ long set_gdt(struct exec_domain *ed,
 {
     struct domain *d = ed->domain;
     /* NB. There are 512 8-byte entries per GDT page. */
-    int i = 0, nr_pages = (entries + 511) / 512;
-    struct desc_struct *vgdt;
+    int i, nr_pages = (entries + 511) / 512;
     unsigned long pfn;
 
-    /* Check the first page in the new GDT. */
-    if ( (pfn = frames[0]) >= max_page )
-        goto fail;
-
+    if ( entries > FIRST_RESERVED_GDT_ENTRY )
+        return -EINVAL;
+    
     shadow_sync_all(d);
 
-    /* The first page is special because Xen owns a range of entries in it. */
-    if ( !get_page_and_type(&frame_table[pfn], d, PGT_gdt_page) )
-    {
-        /* GDT checks failed: try zapping the Xen reserved entries. */
-        if ( !get_page_and_type(&frame_table[pfn], d, PGT_writable_page) )
-            goto fail;
-        vgdt = map_domain_mem(pfn << PAGE_SHIFT);
-        memset(vgdt + FIRST_RESERVED_GDT_ENTRY, 0,
-               NR_RESERVED_GDT_ENTRIES*8);
-        unmap_domain_mem(vgdt);
-        put_page_and_type(&frame_table[pfn]);
-
-        /* Okay, we zapped the entries. Now try the GDT checks again. */
-        if ( !get_page_and_type(&frame_table[pfn], d, PGT_gdt_page) )
-            goto fail;
-    }
-
-    /* Check the remaining pages in the new GDT. */
-    for ( i = 1; i < nr_pages; i++ )
+    /* Check the pages in the new GDT. */
+    for ( i = 0; i < nr_pages; i++ )
         if ( ((pfn = frames[i]) >= max_page) ||
              !get_page_and_type(&frame_table[pfn], d, PGT_gdt_page) )
             goto fail;
-
-    /* Copy reserved GDT entries to the new GDT. */
-    vgdt = map_domain_mem(frames[0] << PAGE_SHIFT);
-    memcpy(vgdt + FIRST_RESERVED_GDT_ENTRY, 
-           gdt_table + FIRST_RESERVED_GDT_ENTRY, 
-           NR_RESERVED_GDT_ENTRIES*8);
-    unmap_domain_mem(vgdt);
 
     /* Tear down the old GDT. */
     destroy_gdt(ed);
 
     /* Install the new GDT. */
+    ed->arch.guest_context.gdt_ents = entries;
     for ( i = 0; i < nr_pages; i++ )
+    {
+        ed->arch.guest_context.gdt_frames[i] = frames[i];
         ed->arch.perdomain_ptes[i] =
             l1e_create_pfn(frames[i], __PAGE_HYPERVISOR);
-
-    SET_GDT_ADDRESS(ed, GDT_VIRT_START(ed));
-    SET_GDT_ENTRIES(ed, entries);
+    }
 
     return 0;
 
@@ -2325,19 +2306,13 @@ long do_set_gdt(unsigned long *frame_list, unsigned int entries)
     unsigned long frames[16];
     long ret;
 
-    if ( (entries <= LAST_RESERVED_GDT_ENTRY) || (entries > 8192) ) 
-        return -EINVAL;
-    
     if ( copy_from_user(frames, frame_list, nr_pages * sizeof(unsigned long)) )
         return -EFAULT;
 
     LOCK_BIGLOCK(current->domain);
 
     if ( (ret = set_gdt(current, frames, entries)) == 0 )
-    {
         local_flush_tlb();
-        __asm__ __volatile__ ("lgdt %0" : "=m" (*current->arch.gdt));
-    }
 
     UNLOCK_BIGLOCK(current->domain);
 
@@ -2352,7 +2327,6 @@ long do_update_descriptor(unsigned long pa, u64 desc)
     unsigned long mfn;
     struct desc_struct *gdt_pent, d;
     struct pfn_info *page;
-    struct exec_domain *ed;
     long ret = -EINVAL;
 
     *(u64 *)&d = desc;
@@ -2379,13 +2353,6 @@ long do_update_descriptor(unsigned long pa, u64 desc)
     switch ( page->u.inuse.type_info & PGT_type_mask )
     {
     case PGT_gdt_page:
-        /* Disallow updates of Xen-reserved descriptors in the current GDT. */
-        for_each_exec_domain(dom, ed) {
-            if ( (l1e_get_pfn(ed->arch.perdomain_ptes[0]) == mfn) &&
-                 (((pa&(PAGE_SIZE-1))>>3) >= FIRST_RESERVED_GDT_ENTRY) &&
-                 (((pa&(PAGE_SIZE-1))>>3) <= LAST_RESERVED_GDT_ENTRY) )
-                goto out;
-        }
         if ( unlikely(!get_page_type(page, PGT_gdt_page)) )
             goto out;
         break;
@@ -2452,7 +2419,7 @@ int revalidate_l1(struct domain *d, l1_pgentry_t *l1page, l1_pgentry_t *snapshot
     int modified = 0, i;
 
 #if 0
-    if ( d->id )
+    if ( d->domain_id )
         printk("%s: l1page mfn=%lx snapshot mfn=%lx\n", __func__,
                l1e_get_pfn(linear_pg_table[l1_linear_offset((unsigned long)l1page)]),
                l1e_get_pfn(linear_pg_table[l1_linear_offset((unsigned long)snapshot)]));
@@ -2836,7 +2803,7 @@ int ptwr_do_page_fault(struct domain *d, unsigned long addr)
     return EXCRET_fault_fixed;
 
  emulate:
-    if ( x86_emulate_memop(get_cpu_user_regs(), addr,
+    if ( x86_emulate_memop(guest_cpu_user_regs(), addr,
                            &ptwr_mem_emulator, BITS_PER_LONG/8) )
         return 0;
     perfc_incrc(ptwr_emulations);
@@ -2870,101 +2837,123 @@ void ptwr_destroy(struct domain *d)
     free_xenheap_page((unsigned long)d->arch.ptwr[PTWR_PT_INACTIVE].page);
 }
 
+int map_pages_to_xen(
+    unsigned long virt,
+    unsigned long pfn,
+    unsigned long nr_pfns,
+    unsigned long flags)
+{
+    l2_pgentry_t *pl2e, ol2e;
+    l1_pgentry_t *pl1e, ol1e;
+    unsigned int  i;
 
+    unsigned int  map_small_pages = !!(flags & MAP_SMALL_PAGES);
+    flags &= ~MAP_SMALL_PAGES;
 
-/************************************************************************/
-/************************************************************************/
-/************************************************************************/
+    while ( nr_pfns != 0 )
+    {
+        pl2e = virt_to_xen_l2e(virt);
 
-/* Graveyard: stuff below may be useful in future. */
-#if 0
-    case MMUEXT_TRANSFER_PAGE:
-        domid  = (domid_t)(val >> 16);
-        gntref = (grant_ref_t)((val & 0xFF00) | ((ptr >> 2) & 0x00FF));
-        
-        if ( unlikely(IS_XEN_HEAP_FRAME(page)) ||
-             unlikely(!pfn_valid(pfn)) ||
-             unlikely((e = find_domain_by_id(domid)) == NULL) )
+        if ( ((((virt>>PAGE_SHIFT) | pfn) & ((1<<PAGETABLE_ORDER)-1)) == 0) &&
+             (nr_pfns >= (1<<PAGETABLE_ORDER)) &&
+             !map_small_pages )
         {
-            MEM_LOG("Bad frame (%p) or bad domid (%d).\n", pfn, domid);
-            okay = 0;
-            break;
-        }
+            /* Super-page mapping. */
+            ol2e  = *pl2e;
+            *pl2e = l2e_create_pfn(pfn, flags|_PAGE_PSE);
 
-        spin_lock(&d->page_alloc_lock);
-
-        /*
-         * The tricky bit: atomically release ownership while there is just one
-         * benign reference to the page (PGC_allocated). If that reference
-         * disappears then the deallocation routine will safely spin.
-         */
-        _d  = pickle_domptr(d);
-        _nd = page->u.inuse._domain;
-        y   = page->count_info;
-        do {
-            x = y;
-            if ( unlikely((x & (PGC_count_mask|PGC_allocated)) != 
-                          (1|PGC_allocated)) ||
-                 unlikely(_nd != _d) )
+            if ( (l2e_get_flags(ol2e) & _PAGE_PRESENT) )
             {
-                MEM_LOG("Bad page values %p: ed=%p(%u), sd=%p,"
-                        " caf=%08x, taf=%08x\n", page_to_pfn(page),
-                        d, d->id, unpickle_domptr(_nd), x, 
-                        page->u.inuse.type_info);
-                spin_unlock(&d->page_alloc_lock);
-                put_domain(e);
-                return 0;
+                local_flush_tlb_pge();
+                if ( !(l2e_get_flags(ol2e) & _PAGE_PSE) )
+                    free_xen_pagetable(l2e_get_page(*pl2e));
             }
-            __asm__ __volatile__(
-                LOCK_PREFIX "cmpxchg8b %2"
-                : "=d" (_nd), "=a" (y),
-                "=m" (*(volatile u64 *)(&page->count_info))
-                : "0" (_d), "1" (x), "c" (NULL), "b" (x) );
-        } 
-        while ( unlikely(_nd != _d) || unlikely(y != x) );
 
-        /*
-         * Unlink from 'd'. At least one reference remains (now anonymous), so
-         * noone else is spinning to try to delete this page from 'd'.
-         */
-        d->tot_pages--;
-        list_del(&page->list);
-        
-        spin_unlock(&d->page_alloc_lock);
-
-        spin_lock(&e->page_alloc_lock);
-
-        /*
-         * Check that 'e' will accept the page and has reservation headroom.
-         * Also, a domain mustn't have PGC_allocated pages when it is dying.
-         */
-        ASSERT(e->tot_pages <= e->max_pages);
-        if ( unlikely(test_bit(DF_DYING, &e->flags)) ||
-             unlikely(e->tot_pages == e->max_pages) ||
-             unlikely(!gnttab_prepare_for_transfer(e, d, gntref)) )
-        {
-            MEM_LOG("Transferee has no reservation headroom (%d,%d), or "
-                    "provided a bad grant ref, or is dying (%p).\n",
-                    e->tot_pages, e->max_pages, e->flags);
-            spin_unlock(&e->page_alloc_lock);
-            put_domain(e);
-            okay = 0;
-            break;
+            virt    += 1UL << L2_PAGETABLE_SHIFT;
+            pfn     += 1UL << PAGETABLE_ORDER;
+            nr_pfns -= 1UL << PAGETABLE_ORDER;
         }
+        else
+        {
+            /* Normal page mapping. */
+            if ( !(l2e_get_flags(*pl2e) & _PAGE_PRESENT) )
+            {
+                pl1e = page_to_virt(alloc_xen_pagetable());
+                clear_page(pl1e);
+                *pl2e = l2e_create_phys(__pa(pl1e), __PAGE_HYPERVISOR);
+            }
+            else if ( l2e_get_flags(*pl2e) & _PAGE_PSE )
+            {
+                pl1e = page_to_virt(alloc_xen_pagetable());
+                for ( i = 0; i < L1_PAGETABLE_ENTRIES; i++ )
+                    pl1e[i] = l1e_create_pfn(
+                        l2e_get_pfn(*pl2e) + i,
+                        l2e_get_flags(*pl2e) & ~_PAGE_PSE);
+                *pl2e = l2e_create_phys(__pa(pl1e), __PAGE_HYPERVISOR);
+                local_flush_tlb_pge();
+            }
 
-        /* Okay, add the page to 'e'. */
-        if ( unlikely(e->tot_pages++ == 0) )
-            get_knownalive_domain(e);
-        list_add_tail(&page->list, &e->page_list);
-        page_set_owner(page, e);
+            pl1e  = l2e_to_l1e(*pl2e) + l1_table_offset(virt);
+            ol1e  = *pl1e;
+            *pl1e = l1e_create_pfn(pfn, flags);
+            if ( (l1e_get_flags(ol1e) & _PAGE_PRESENT) )
+                local_flush_tlb_one(virt);
 
-        spin_unlock(&e->page_alloc_lock);
+            virt    += 1UL << L1_PAGETABLE_SHIFT;
+            pfn     += 1UL;
+            nr_pfns -= 1UL;
+        }
+    }
 
-        /* Transfer is all done: tell the guest about its new page frame. */
-        gnttab_notify_transfer(e, d, gntref, pfn);
-        
-        put_domain(e);
-        break;
+    return 0;
+}
+
+void __set_fixmap(
+    enum fixed_addresses idx, unsigned long p, unsigned long flags)
+{
+    if ( unlikely(idx >= __end_of_fixed_addresses) )
+        BUG();
+    map_pages_to_xen(fix_to_virt(idx), p >> PAGE_SHIFT, 1, flags);
+}
+
+#ifdef MEMORY_GUARD
+
+void memguard_init(void)
+{
+    map_pages_to_xen(
+        PAGE_OFFSET, 0, xenheap_phys_end >> PAGE_SHIFT,
+        __PAGE_HYPERVISOR|MAP_SMALL_PAGES);
+}
+
+static void __memguard_change_range(void *p, unsigned long l, int guard)
+{
+    unsigned long _p = (unsigned long)p;
+    unsigned long _l = (unsigned long)l;
+    unsigned long flags = __PAGE_HYPERVISOR | MAP_SMALL_PAGES;
+
+    /* Ensure we are dealing with a page-aligned whole number of pages. */
+    ASSERT((_p&PAGE_MASK) != 0);
+    ASSERT((_l&PAGE_MASK) != 0);
+    ASSERT((_p&~PAGE_MASK) == 0);
+    ASSERT((_l&~PAGE_MASK) == 0);
+
+    if ( guard )
+        flags &= ~_PAGE_PRESENT;
+
+    map_pages_to_xen(
+        _p, virt_to_phys(p) >> PAGE_SHIFT, _l >> PAGE_SHIFT, flags);
+}
+
+void memguard_guard_range(void *p, unsigned long l)
+{
+    __memguard_change_range(p, l, 1);
+}
+
+void memguard_unguard_range(void *p, unsigned long l)
+{
+    __memguard_change_range(p, l, 0);
+}
+
 #endif
 
 /*
